pyproject.toml
[project]
name = "ai6130-groupproject"
version = "0.1.0"
description = "HotpotQA RAG experimentation framework built with LangChain"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.8.4",
    "faiss-cpu>=1.8.0",
    "ipykernel>=7.2.0",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.1",
    "numpy>=2.2.0",
    "pandas>=2.2.3",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=3.4.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
]

[project.scripts]
hotpot-rag = "hotpot_rag.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["hotpot_rag"]

[tool.pytest.ini_options]
testpaths = ["tests"]
